pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iPhoto"
version = "0.1.0"
description = "Folder-native photo manager with Live Photo support"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "iPhoto Team"}]
dependencies = [
  "typer>=0.12",
  "rich>=13",
  "jsonschema>=4",
  "Pillow>=10",
  "pillow-heif>=0.16",
  "imagehash>=4",
  "xxhash>=3",
  "python-dateutil>=2",
]

[project.scripts]
iphoto = "iPhoto.cli:app"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "N", "ASYNC", "S", "BLE", "A", "C4", "PIE", "Q", "RUF"]
ignore = ["E203"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
pythonpath = ["src"]

[tool.typer]
pretty_exceptions_enable = true
